refactor - emoji list from usecase
